+++ src/components/Geosearch.vue
@@ -7,7 +7,6 @@ import axios from "axios";
 import Plotly from "plotly.js";
 
 let graphParameterCodeList = "00065,63160,72279";
-let timeQueryRange = "&period=P7D";
 
 export default {
   props: ["map"],
@@ -37,7 +36,7 @@ export default {
         siteCode +
         "</br>" +
         siteName +
-        '</label></br><p id="graphLoadMessage"><v-progress-circular indeterminate :width=3 :size=20></v-progress-circular><span> NWIS data graph loading...</span></p><div id="graphContainer" style="width:100%; height:200px;display:none;"></div> <div id="dataCredit">Gage Height data courtesy of the U.S. Geological Survey</div><a class="nwis-link" target="_blank" href="https://nwis.waterdata.usgs.gov/nwis/uv?site_no=' +
+        '</label></br><p id="graphLoadMessage"><v-progress-circular indeterminate :width=3 :size=20></v-progress-circular><span> NWIS data graph loading...</span></p><div id="graphContainer" style="width:100%; min-height:200px;display:none;"></div> <div id="dataCredit">Gage Height data courtesy of the U.S. Geological Survey</div><a class="nwis-link" target="_blank" href="https://nwis.waterdata.usgs.gov/nwis/uv?site_no=' +
         siteCode +
         '"><b>Site ' +
         siteCode +
@@ -47,7 +46,7 @@ export default {
         siteCode +
         "&parameterCd=" +
         graphParameterCodeList +
-        timeQueryRange;
+        this.timePeriodValue;
       axios.get(url).then((data) => {
         if (
           data.data == undefined ||
@@ -248,6 +247,13 @@ export default {
       };
     };
   },
+  computed: {
+    timePeriodValue: {
+      get() {
+        return this.$store.state.selectedTimePeriodState;
+      },
+    },
+  },
 };
 </script>
 
@@ -297,7 +303,8 @@ input:not([disabled]):focus {
 }
 
 @media screen and (max-width: 1024px) {
-.v-toolbar__content, .v-toolbar__extension {
+  .v-toolbar__content,
+  .v-toolbar__extension {
     padding: 4px 8px 4px 16px !important;
   }
 }
@@ -335,10 +342,6 @@ input:not([disabled]):focus {
   .v-toolbar__extension {
     padding: 4px 8px 4px 8px !important;
   }
-}
-
-.leaflet-popup-content {
-  margin: -1px 10px 10px 10px !important;
 }
 
 #dataCredit {