+++ CHANGELOG.md
@@ -19,6 +19,27 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 -
 
+## [v0.6.0](https://github.com/USGS-WiM/Thresholds/releases/tag/v0.6.0) - 9/22/2021
+
+### Added
+
+- Add tooltips for active flooding
+- Added NOAA Tides and Current Stations supporting layer
+- Added ability to change time period
+- Added alert with thresholds exceeded count to sidebar
+
+### Changed
+
+- Reposition map components on mobile
+- Disable "Show All Active Flooding" button when RP layers are all checked
+- Changed "Show All Active Flooding" button text to "Show All Active Flooding Layers"
+
+### Fixed
+
+- Fix lat/long/zoom display
+- position of popups
+- Fixed layer order when toggling RP layer on/off
+
 ## [v0.5.0](https://github.com/USGS-WiM/Thresholds/releases/tag/v0.5.0) - 9/1/2021
 
 ### Added